Finalized show_bounds and separated online_callbacks into smaller files
