+++ bioptim/__init__.py
@@ -231,4 +231,4 @@ from .optimization.stochastic_optimal_control_program import StochasticOptimalCo
 from .optimization.problem_type import SocpType
 from .misc.casadi_expand import lt, le, gt, ge, if_else, if_else_zero
 
-from .gui.online_callback import OnlineCallbackServer
+from .gui.online_callback_server import OnlineCallbackServerBackend
