+++ bioptim/gui/online_callback_server.py
@@ -1,285 +1,17 @@
-from abc import ABC, abstractmethod
 from enum import Enum
 import json
 import logging
-import multiprocessing as mp
 import socket
 import struct
 import threading
 
-from casadi import Callback, nlpsol_out, nlpsol_n_out, Sparsity, DM
+from casadi import nlpsol_out, DM
 from matplotlib import pyplot as plt
 import numpy as np
 
+from .online_callback_abstract import OnlineCallbackAbstract
 from .plot import PlotOcp, OcpSerializable
 from ..optimization.optimization_vector import OptimizationVectorHelper
-
-
-class OnlineCallbackAbstract(Callback, ABC):
-    """
-    CasADi interface of Ipopt callbacks
-
-    Attributes
-    ----------
-    ocp: OptimalControlProgram
-        A reference to the ocp to show
-    nx: int
-        The number of optimization variables
-    ng: int
-        The number of constraints
-
-    Methods
-    -------
-    get_n_in() -> int
-        Get the number of variables in
-    get_n_out() -> int
-        Get the number of variables out
-    get_name_in(i: int) -> int
-        Get the name of a variable
-    get_name_out(_) -> str
-        Get the name of the output variable
-    get_sparsity_in(self, i: int) -> tuple[int]
-        Get the sparsity of a specific variable
-    eval(self, arg: list | tuple, force: bool = False) -> list[int]
-        Send the current data to the plotter
-    """
-
-    def __init__(self, ocp, opts: dict = None, show_options: dict = None):
-        """
-        Parameters
-        ----------
-        ocp: OptimalControlProgram
-            A reference to the ocp to show
-        opts: dict
-            Option to AnimateCallback method of CasADi
-        show_options: dict
-            The options to pass to PlotOcp
-        """
-        if opts is None:
-            opts = {}
-
-        Callback.__init__(self)
-        self.ocp = ocp
-        self.nx = self.ocp.variables_vector.shape[0]
-
-        # There must be an option to add an if here
-        from ..interfaces.ipopt_interface import IpoptInterface
-
-        interface = IpoptInterface(ocp)
-        all_g, _ = interface.dispatch_bounds()
-        self.ng = all_g.shape[0]
-
-        self.construct("AnimateCallback", opts)
-
-    @abstractmethod
-    def close(self):
-        """
-        Close the callback
-        """
-
-    @staticmethod
-    def get_n_in() -> int:
-        """
-        Get the number of variables in
-
-        Returns
-        -------
-        The number of variables in
-        """
-
-        return nlpsol_n_out()
-
-    @staticmethod
-    def get_n_out() -> int:
-        """
-        Get the number of variables out
-
-        Returns
-        -------
-        The number of variables out
-        """
-
-        return 1
-
-    @staticmethod
-    def get_name_in(i: int) -> int:
-        """
-        Get the name of a variable
-
-        Parameters
-        ----------
-        i: int
-            The index of the variable
-
-        Returns
-        -------
-        The name of the variable
-        """
-
-        return nlpsol_out(i)
-
-    @staticmethod
-    def get_name_out(_) -> str:
-        """
-        Get the name of the output variable
-
-        Returns
-        -------
-        The name of the output variable
-        """
-
-        return "ret"
-
-    def get_sparsity_in(self, i: int) -> tuple:
-        """
-        Get the sparsity of a specific variable
-
-        Parameters
-        ----------
-        i: int
-            The index of the variable
-
-        Returns
-        -------
-        The sparsity of the variable
-        """
-
-        n = nlpsol_out(i)
-        if n == "f":
-            return Sparsity.scalar()
-        elif n in ("x", "lam_x"):
-            return Sparsity.dense(self.nx)
-        elif n in ("g", "lam_g"):
-            return Sparsity.dense(self.ng)
-        else:
-            return Sparsity(0, 0)
-
-    @abstractmethod
-    def eval(self, arg: list | tuple, force: bool = False) -> list:
-        """
-        Send the current data to the plotter
-
-        Parameters
-        ----------
-        arg: list | tuple
-            The data to send
-
-        Returns
-        -------
-        A list of error index
-        """
-
-
-class OnlineCallbackMultiprocess(OnlineCallbackAbstract):
-    """
-    Multiprocessing implementation of the online callback
-
-    Attributes
-    ----------
-    queue: mp.Queue
-        The multiprocessing queue
-    plotter: ProcessPlotter
-        The callback for plotting for the multiprocessing
-    plot_process: mp.Process
-        The multiprocessing placeholder
-    """
-
-    def __init__(self, ocp, opts: dict = None, show_options: dict = None):
-        super(OnlineCallbackMultiprocess, self).__init__(ocp, opts, show_options)
-
-        self.queue = mp.Queue()
-        self.plotter = self.ProcessPlotter(self.ocp)
-        self.plot_process = mp.Process(target=self.plotter, args=(self.queue, show_options), daemon=True)
-        self.plot_process.start()
-
-    def close(self):
-        self.plot_process.kill()
-
-    def eval(self, arg: list | tuple, force: bool = False) -> list:
-        # Dequeuing the data by removing previous not useful data
-        while not self.queue.empty():
-            self.queue.get_nowait()
-
-        args_dict = {}
-        for i, s in enumerate(nlpsol_out()):
-            args_dict[s] = arg[i]
-        self.queue.put_nowait(args_dict)
-        return [0]
-
-    class ProcessPlotter(object):
-        """
-        The plotter that interface PlotOcp and the multiprocessing
-
-        Attributes
-        ----------
-        ocp: OptimalControlProgram
-            A reference to the ocp to show
-        pipe: mp.Queue
-            The multiprocessing queue to evaluate
-        plot: PlotOcp
-            The handler on all the figures
-
-        Methods
-        -------
-        callback(self) -> bool
-            The callback to update the graphs
-        """
-
-        def __init__(self, ocp):
-            """
-            Parameters
-            ----------
-            ocp: OptimalControlProgram
-                A reference to the ocp to show
-            """
-
-            self._ocp: OcpSerializable = ocp
-            self._plotter: PlotOcp = None
-            self._update_time = 0.001
-
-        def __call__(self, pipe: mp.Queue, show_options: dict | None):
-            """
-            Parameters
-            ----------
-            pipe: mp.Queue
-                The multiprocessing queue to evaluate
-            show_options: dict
-                The option to pass to PlotOcp
-            """
-
-            if show_options is None:
-                show_options = {}
-            self._pipe = pipe
-
-            dummy_phase_times = OptimizationVectorHelper.extract_step_times(self._ocp, DM(np.ones(self._ocp.n_phases)))
-            self._plotter = PlotOcp(self._ocp, dummy_phase_times=dummy_phase_times, **show_options)
-            threading.Timer(self._update_time, self.plot_update).start()
-            plt.show()
-
-        def plot_update(self) -> bool:
-            """
-            The callback to update the graphs
-
-            Returns
-            -------
-            True if everything went well
-            """
-
-            args = {}
-            while not self._pipe.empty():
-                args = self._pipe.get()
-
-            if args:
-                self._plotter.update_data(*self._plotter.parse_data(**args), **args)
-
-            # We want to redraw here to actually consume a bit of time, otherwise it goes to fast and pipe remains empty
-            for fig in self._plotter.all_figures:
-                fig.canvas.draw()
-            if [plt.fignum_exists(fig.number) for fig in self._plotter.all_figures].count(True) > 0:
-                # If there are still figures, we keep updating
-                threading.Timer(self._update_time, self.plot_update).start()
-
-            return True
 
 
 _default_host = "localhost"
@@ -294,15 +26,16 @@ def _deserialize_show_options(show_options: bytes) -> dict:
     return json.loads(show_options.decode())
 
 
-class OnlineCallbackServer:
-    class _ServerMessages(Enum):
-        INITIATE_CONNEXION = 0
-        NEW_DATA = 1
-        CLOSE_CONNEXION = 2
-        EMPTY = 3
-        TOO_SOON = 4
-        UNKNOWN = 5
+class _ServerMessages(Enum):
+    INITIATE_CONNEXION = 0
+    NEW_DATA = 1
+    CLOSE_CONNEXION = 2
+    EMPTY = 3
+    TOO_SOON = 4
+    UNKNOWN = 5
 
+
+class OnlineCallbackServerBackend:
     def _prepare_logger(self):
         name = "OnlineCallbackServer"
         console_handler = logging.StreamHandler()
@@ -351,15 +84,15 @@ class OnlineCallbackServer:
             self._logger.debug("Waiting for data from client")
             data = client_socket.recv(1024)
             if not data:
-                return OnlineCallbackServer._ServerMessages.EMPTY, None
+                return _ServerMessages.EMPTY, None
         except:
             self._logger.warning("Client closed connexion")
             client_socket.close()
-            return OnlineCallbackServer._ServerMessages.CLOSE_CONNEXION, None
+            return _ServerMessages.CLOSE_CONNEXION, None
 
         data_as_list = data.decode().split("\n")
         try:
-            message_type = OnlineCallbackServer._ServerMessages(int(data_as_list[0]))
+            message_type = _ServerMessages(int(data_as_list[0]))
             len_all_data = [int(len_data) for len_data in data_as_list[1][1:-1].split(",")]
             # Sends confirmation and waits for the next message
             if send_confirmation:
@@ -374,23 +107,23 @@ class OnlineCallbackServer:
                 client_socket.sendall("OK".encode())
         except ValueError:
             self._logger.warning("Unknown message type received")
-            message_type = OnlineCallbackServer._ServerMessages.UNKNOWN
+            message_type = _ServerMessages.UNKNOWN
             # Sends failure
             if send_confirmation:
                 client_socket.sendall("NOK".encode())
             data_out = []
 
-        if message_type == OnlineCallbackServer._ServerMessages.CLOSE_CONNEXION:
+        if message_type == _ServerMessages.CLOSE_CONNEXION:
             self._logger.info("Received close connexion from client")
             client_socket.close()
             plt.close()
-            return OnlineCallbackServer._ServerMessages.CLOSE_CONNEXION, None
+            return _ServerMessages.CLOSE_CONNEXION, None
 
         return message_type, data_out
 
     def _wait_for_new_connexion(self, client_socket: socket.socket):
         message_type, data = self._wait_for_data(client_socket=client_socket)
-        if message_type == OnlineCallbackServer._ServerMessages.INITIATE_CONNEXION:
+        if message_type == _ServerMessages.INITIATE_CONNEXION:
             self._logger.debug(f"Received hand shake from client")
             self._initialize_plotter(client_socket, data)
 
@@ -451,7 +184,7 @@ class OnlineCallbackServer:
 
         should_continue = False
         message_type, data = self._wait_for_data(client_socket=client_socket, send_confirmation=False)
-        if message_type == OnlineCallbackServer._ServerMessages.NEW_DATA:
+        if message_type == _ServerMessages.NEW_DATA:
             try:
                 self._update_data(data)
                 should_continue = True
@@ -459,10 +192,7 @@ class OnlineCallbackServer:
                 self._logger.warning("Error while updating data from client, closing connexion")
                 plt.close()
                 client_socket.close()
-        elif (
-            message_type == OnlineCallbackServer._ServerMessages.EMPTY
-            or message_type == OnlineCallbackServer._ServerMessages.CLOSE_CONNEXION
-        ):
+        elif message_type == _ServerMessages.EMPTY or message_type == _ServerMessages.CLOSE_CONNEXION:
             self._logger.debug("Received empty data from client (end of stream), closing connexion")
 
         if should_continue:
@@ -556,7 +286,7 @@ class OnlineCallbackServer(OnlineCallbackAbstract):
 
         # Sends message type and dimensions
         self._socket.sendall(
-            f"{OnlineCallbackServer._ServerMessages.INITIATE_CONNEXION.value}\n{[len(serialized_ocp), len(serialized_show_options)]}".encode()
+            f"{_ServerMessages.INITIATE_CONNEXION.value}\n{[len(serialized_ocp), len(serialized_show_options)]}".encode()
         )
         if self._socket.recv(1024).decode() != "OK":
             raise RuntimeError("The server did not acknowledge the connexion")
@@ -577,9 +307,7 @@ class OnlineCallbackServer(OnlineCallbackAbstract):
         )
 
     def close(self):
-        self._socket.sendall(
-            f"{OnlineCallbackServer._ServerMessages.CLOSE_CONNEXION.value}\nGoodbye from client!".encode()
-        )
+        self._socket.sendall(f"{_ServerMessages.CLOSE_CONNEXION.value}\nGoodbye from client!".encode())
         self._socket.close()
 
     def eval(self, arg: list | tuple, force: bool = False) -> list:
@@ -623,9 +351,7 @@ class OnlineCallbackServer(OnlineCallbackAbstract):
                 y_steps_tp = y_steps.tolist()
                 data_serialized += struct.pack("d" * len(y_steps_tp), *y_steps_tp)
 
-        self._socket.sendall(
-            f"{OnlineCallbackServer._ServerMessages.NEW_DATA.value}\n{[len(header), len(data_serialized)]}".encode()
-        )
+        self._socket.sendall(f"{_ServerMessages.NEW_DATA.value}\n{[len(header), len(data_serialized)]}".encode())
         # If send_confirmation is True, we should wait for the server to acknowledge the data here (sends OK)
         self._socket.sendall(header.encode())
         self._socket.sendall(data_serialized)
